requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
requests>=2.31.0
urllib3>=2.0.0
flask>=2.0.0
aiohttp

